--- a/app/controllers/org_admin/plans_controller.rb
+++ b/app/controllers/org_admin/plans_controller.rb
@@ -17,7 +17,11 @@ module OrgAdmin
 
       @super_admin = current_user.can_super_admin?
       @clicked_through = params[:click_through].present?
-      @plans = @super_admin ? Plan.all.page(1).includes(:template, roles: { user: :org }) : current_user.org.org_admin_plans.page(1)
+      @plans = if @super_admin
+                 Plan.all.page(1).includes(:template, roles: { user: :org })
+               else
+                 current_user.org.org_admin_plans.page(1)
+               end
     end
     # rubocop:enable Metrics/AbcSize