Changes proposed in this PR:
- Address bullet warnings for the following paths:
  - `/super_admin/api_clients`
  - `/public_plans`
  - `org_admins/plans` (Superuser only)

The following table compares the `development` branch to `aaron/bullet-fixes` by making requests to the various paths affected by this PR. The benchmarking was performed via ab - Apache HTTP server benchmarking tool alongside an older (May 2024) db dump from the production environment of DMP Assistant. In both cases, 100 consecutive requests were performed to each path and the recorded result is the mean request time (ms).

Example request:
```
$ ab -n 100 -k -C "_dmp_roadmap_session=COOKIE_VALUE" -l http://127.0.0.1:3000/PATH/TO/PAGE
```
Path | mean request time (ms) before | mean request time (ms) now | % improvement
-- | -- | -- | --
/super_admin/api_clients | 2901 | 843 | 70.9
/public_plans | 232 | 168 | 27.6
/org_admin/plans (as super_admin type user) | 269 | 190 | 29.4